+++ docs/source/release_notes/release-notes.rst
@@ -99,6 +99,10 @@ develop
          - TimeLock Server now logs that a new client has been registered the first time a service makes a request (for each lifetime of each server).
            (`Pull Request <https://github.com/palantir/atlasdb/pull/3676>`__)
 
+    *    - |new|
+         - Adds a debug/statistics gathering CLI useful for users looking to rewrite their schema to work better with targeted sweep
+           (`Pull Request <https://github.com/palantir/atlasdb/pull/3710>`__)
+
 ========
 v0.114.0
 ========